test: suppress torchvision Pillow deprecation warning
Added filter to ignore DeprecationWarning from torchvision about
Pillow's 'mode' parameter. This is a known issue in torchvision
that will be fixed when they update for Pillow 13 (2026).

The warning is:
  'mode' parameter is deprecated and will be removed in Pillow 13

This is external to our code and doesn't affect functionality.
Suppressing it keeps test output clean and focused on actual issues.